fix: update configuration and refactor order-related classes for consistency

+++ services/order/src/main/java/com/rares/ecommerce/order/Integration/Client/Product/DTO/PurchaseResponse.java
@@ -7,6 +7,6 @@ public record PurchaseResponse(
         String name,
         String description,
         BigDecimal price,
-        double quantity
+        Integer quantity
 ) {
 }
