+++ services/order/src/main/java/com/rares/ecommerce/order/OrderLine/DTO/OrderLineRequest.java
@@ -4,5 +4,5 @@ public record OrderLineRequest(
         Integer id,
         Integer orderId,
         Integer productId,
-        double quantity) {
+        Integer quantity) {
 }
